pipe/phono-deps: move phono.SingleUse to pipe
