--- phono.go
@@ -1,20 +1,0 @@
-package phono
-
-import (
-	"errors"
-	"sync"
-)
-
-// SingleUse is designed to be used in runner-return functions to define a single-use pipe components.
-func SingleUse(once *sync.Once) (err error) {
-	err = ErrSingleUseReused
-	once.Do(func() {
-		err = nil
-	})
-	return
-}
-
-var (
-	// ErrSingleUseReused is returned when object designed for single-use is being reused.
-	ErrSingleUseReused = errors.New("Error reuse single-use object")
-)
